+++ src/main/java/insane96mcp/iguanatweaksreborn/module/combat/stats/Stats.java
@@ -59,7 +59,7 @@ public class Stats extends Feature {
 	@Label(name = "Decimal arrow's damage", description = "If true, arrows will deal float damage instead of being rounded up.")
 	public static Boolean decimalArrowsDamage = true;
 	@Config(min = 0d, max = 10d)
-	@Label(name = "Arrows Damage Multiplier", description = "Multiplies arrow's damage if shot from bow by this value. (this doesn't affect mobs arrows)")
+	@Label(name = "Arrows Damage Multiplier", description = "Multiplies arrow's damage by this value. (this doesn't affect mobs arrows)")
 	public static Double arrowsDamageMultiplier = 0.5d;
 	@Config(min = 0d, max = 10d)
 	@Label(name = "Crossbow velocity", description = "Speed at which arrows are shot from crossbows. Vanilla is 3.15")