+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Upcoming
+* Fixed mobs arrows being affected by arrows damage multiplier
+
 ## 3.13.2
 * Smite now applies 1s (plus up to 0.5s per level) of Weakness II to undead creatures
 * Increased wooden tools durability